% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateIndexes.R
\name{checkAnyResampleIndexesExists}
\alias{checkAnyResampleIndexesExists}
\title{Check if any of the resample indexes elements exist}
\usage{
checkAnyResampleIndexesExists(folder_path)
}
\description{
Check if any of the resample indexes elements exist
}

