% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{checkCalibration}
\alias{checkCalibration}
\title{Build calibration plots for a \code{multiClass} object}
\usage{
checkCalibration(fitted_model, resample_indexes, number_bins,
  order_smooth = 2, replicate_indexes = NULL)
}
\description{
Build calibration plots for a \code{multiClass} object
}

