\name{evaluateProbClass.multiClass}
\alias{evaluateProbClass.multiClass}
\title{Evaluate predicted class probabilities for a given score function.}
\usage{
\method{evaluateProbClass}{multiClass}(pred_obj, resample_indexes, type,
  ignore_tag = FALSE, ...)
}
\arguments{
  \item{pred_obj}{A \code{multiClass} object, usually
  generated from the \code{datasetResample} object with
  some predictive model with an interface function to the
  \code{tgmMultiClass} package.}

  \item{resample_indexes}{A \code{datasetResample} object,
  usually generated by the \code{generateTestIndexes}
  function.}

  \item{type}{Which type of score function should be used.
  Currently, on \code{log_score} is available.}

  \item{ignore_tag}{Before computing scores, we check to
  see if \code{pred_obj} was indeed generated from the
  \code{resample_indexes}.  In case of a mismatch an error
  is issued, unless \code{ignore_tag = FALSE}, in which
  case just a warning is issued.}
}
\value{
An S3 object of class \code{multiClassScores}.
}
\description{
Evaluate predicted class probabilities for a given score
function.
}
\details{
Given a \code{datasetResample} object and a compatible
\code{multiClass} prediction object it computes prediction
scores of some \code{type} for each replication.
}

