\name{evaluateProbClass.multiClassValidation}
\alias{evaluateProbClass.multiClassValidation}
\title{Evaluate predicted class probabilities over a validation set for a given score function.}
\usage{
\method{evaluateProbClass}{multiClassValidation}(pred_obj, resample_indexes,
  type, ignore_tag = FALSE, replicate_index = NULL, ...)
}
\arguments{
  \item{pred_obj}{A \code{multiClassValidation} object,
  usually generated from the \code{datasetResample} object
  with some set of predictive models through an interface
  function to the \code{tgmMultiClass} package.}

  \item{resample_indexes}{A \code{datasetResample} object,
  usually generated by the \code{generateTestIndexes}
  function.}

  \item{type}{Which type of score function should be used.
  Currently, on \code{log_score} is available.}

  \item{ignore_tag}{Before computing scores, we check to
  see if \code{pred_obj} was indeed generated from the
  \code{resample_indexes}.  In case of a mismatch an error
  is issued, unless \code{ignore_tag = FALSE}, in which
  case just a warning is issued.}
}
\value{
An S3 object of class \code{multiClassValidationScores}.
}
\description{
Evaluate predicted class probabilities over a validation
set for a given score function.
}
\details{
Given a \code{datasetResample} object and a compatible
\code{multiClassValidation} prediction object it computes
prediction scores of some \code{type} for each model used
in the validation step.
}

