% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateIndexes.R
\name{loadResampleIndexes}
\alias{loadResampleIndexes}
\title{Load a resample_indexes object from txt files.}
\usage{
loadResampleIndexes(folder_path, optional_name = NULL,
  stringsAsFactors = FALSE)
}
\description{
Load a resample_indexes object from txt files.
}

