% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateIndexes.R
\name{mcGet.datasetResample}
\alias{mcGet.datasetResample}
\title{Extract info from \code{datasetResample} objects}
\usage{
\method{mcGet}{datasetResample}(x, attr, i = NULL)
}
\description{
Extract info from \code{datasetResample} objects
}

