% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiClass.R
\name{mcGet.multiClass}
\alias{mcGet.multiClass}
\title{Extract info from \code{multiClass} objects}
\usage{
\method{mcGet}{multiClass}(x, attr, i = NULL)
}
\description{
Extract info from \code{multiClass} objects
}

