% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.R
\name{mcGet.multiClassValidation}
\alias{mcGet.multiClassValidation}
\title{Extract info from \code{multiClassValidation} objects}
\usage{
\method{mcGet}{multiClassValidation}(x, attr, i = NULL)
}
\description{
Extract info from \code{multiClassValidation} objects
}

