% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mcGet.multiClassValidationScores}
\alias{mcGet.multiClassValidationScores}
\title{Extract info from \code{multiClassValidationScores} objects}
\usage{
\method{mcGet}{multiClassValidationScores}(x, attr, i = NULL)
}
\description{
Extract info from \code{multiClassValidationScores} objects
}

