% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mcGet.summarizedValidation}
\alias{mcGet.summarizedValidation}
\title{Get elements from \code{summarizedValidation} objects}
\usage{
\method{mcGet}{summarizedValidation}(x, attr, ...)
}
\description{
Get elements from \code{summarizedValidation} objects
}

