\name{multiClass}
\alias{multiClass}
\title{Constructor for \code{multiClass} objects}
\usage{
multiClass(class_labels, prob, datasetResample_tag, summary_validation = NULL,
  metric_name = NULL)
}
\arguments{
  \item{class_labels}{Character vector with the names of
  the class labels. See details.}

  \item{prob}{List where each element contains a numeric
  vector with predictions for a given test set. The numeric
  vector has the form c(predictions_class1,
  predictions_class2, ..., predictions_classK) for
  classification with K classes.}

  \item{datasetResample_tag}{Character vector representing
  a tag that connects this \code{multiClass} object with a
  data resample object, usually generated by functions like
  \code{generateTestIndexes}}

  \item{summary_validation}{Object of class
  \code{summaryValidation}, usually created by
  \code{SummaryValidationMetric} function. Default to NULL,
  when no validation step is necessary.}

  \item{metric_name}{Character vector indicating the name
  of the metric used in the validation step. Default to
  NULL, when no validation step is necessary. See the help
  of \code{evaluateProbClass} function to which metrics are
  currently available.}
}
\value{
S3 object of class \code{multiClass}.
}
\description{
\code{multiClass} objects hold final test set predictions
performed by models.
}
\details{
\code{multiClass} objects hold predictions performed by a
model over a number of test sets generated according to
some resampling of the data. A tag exist to connect the
predictions with the object that contains information about
the selected test sets. This allows results from different
models to be compared since two \code{multiClass} objects
with the same tag are guaranteed to be generated from the
same test set points.

For developers planning to create an interface for their
models that uses \code{multiClass}: Make sure that the
\code{class_labels} are the same as the one contained in
the \code{datasetResample} object used to generate the
prediction.
}
\seealso{
\code{\link{generateTestIndexes}},
\code{\link{summaryValidation}},
\code{\link{evaluateProbClass}}
}

