% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{multiClassScores}
\alias{multiClassScores}
\title{#Compute prediction log scores for a \code{multiClass} object.
#
#}
\usage{
multiClassScores(scores_list, resample_indexes, score_type)
}
\arguments{
\item{scores_list}{List where each element contains a numerical vector with the score
values for each prediction point in a specific test set.}

\item{resample_indexes}{S3 object of class \code{datasetResample} with information
about Class Labels and the data used using for training and predictions. Usually
generated by the \code{generateTestIndexes} function.}

\item{score_type}{Which type of score function was used.}
}
\value{
A matrix where each column represents the prediction log scores
#for a given replication. The number of rows is equal to the number
#of prediction points.
Constructor for \code{multiClassScores} objects

\code{multiClassScores} objects hold Scores obtained from comparing prediction with test set observations.

\code{multiClassScores} objects hold Scores obtained by comparing predictions performed
over a series of test sets with actual observed values. A tag exist
to connect the scores with the object that contains information about
the selected test sets. This allows results from different models to be
compared since two \code{multiClassScores} objects with the same tag are guaranteed
to be generated from the same test set points.

S3 object of class \code{multiClassScores}.
}
\description{
#Compute prediction log scores for a \code{multiClass} object.
#
#
}
\details{
Given a \code{datasetResample} object and a compatible
#\code{multiClass} prediction object it compute log scores of the
#predictions for each replication.
#
#
}
\seealso{
\code{\link{generateTestIndexes}},
 \code{\link{summaryValidation}},
 \code{\link{evaluateProbClass}}

Other tgmMultiClass.constructors: \code{\link{multiClassValidationScores}};
  \code{\link{multiClassValidation}};
  \code{\link{multiClass}}
}

