\name{multiClassValidation}
\alias{multiClassValidation}
\title{Constructor for \code{multiClassValidation} objects}
\usage{
multiClassValidation(resample_indexes, prob, replicate_index)
}
\arguments{
  \item{resample_indexes}{S3 object of class
  \code{datasetResample} with information about Class
  Labels and the data used using for training and
  predictions. Usually generated by the
  \code{generateTestIndexes} function.}

  \item{prob}{List where each element contains a numeric
  vector with predictions performed by a model on the same
  validation set. The numeric vector has the form
  c(predictions_class1, predictions_class2, ...,
  predictions_classK) for classification with K classes.}

  \item{replicate_index}{Integer number that indicates
  which validation set was used.}
}
\value{
S3 object of class \code{multiClassValidation}.
}
\description{
\code{multiClassValidation} objects hold validation set
predictions performed by different models.
}
\details{
\code{multiClassValidation} objects hold predictions
performed by different models over the same validation set.
A tag exist to connect the predictions with the object that
contains information about the selected validation set. In
addition the specific replicate index of the validation set
used is stored because several validation sets are
contained in the tagged object.
}
\seealso{
\code{\link{generateTestIndexes}},

Other tgmMultiClass.constructors: \code{\link{multiClass}}
}

