\name{multiClassValidationScores}
\alias{multiClassValidationScores}
\title{Constructor for \code{multiClassValidationScores} objects}
\usage{
multiClassValidationScores(scores_list, resample_indexes, replicate_index,
  score_type, tune_grid)
}
\arguments{
  \item{scores_list}{List where each element contains a
  numerical vector with the score values for each
  prediction point generated by a specific model for a
  specific validation set.}

  \item{resample_indexes}{S3 object of class
  \code{datasetResample} with information about Class
  Labels and the data used using for training and
  predictions. Usually generated by the
  \code{generateTestIndexes} function.}

  \item{replicate_index}{Integer number that indicates
  which validation set was used.}

  \item{score_type}{Which type of score function was used.}

  \item{tune_grid}{It is a matrix with named columns
  containing tuning parameters of the model. Each row of
  the matrix represents tuning parameter values of a
  specific model.}
}
\value{
S3 object of class \code{multiClassValidationScores}.
}
\description{
\code{multiClassValidationScores} objects hold Scores
obtained from comparing multiple model predictions to a
specific validation set. A tag and replication index exist
to connect the scores with the object that contains
information about the selected validation set. This allow
us to identify what was the exact validation set used to
produce the scores.
}
\seealso{
\code{\link{generateTestIndexes}},
\code{\link{summaryValidation}},
\code{\link{evaluateProbClass}}

Other tgmMultiClass.constructors:
\code{\link{multiClassScores}};
\code{\link{multiClassValidation}};
\code{\link{multiClass}}
}

