% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{pickBestValidationModel}
\alias{pickBestValidationModel}
\title{Select the best model based on computed metric}
\usage{
pickBestValidationModel(validation_scores, summarize_function = function(x)
  mean(x, na.rm = TRUE))
}
\description{
Select the best model based on computed metric
}

