% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot.summarizedValidation}
\alias{plot.summarizedValidation}
\title{Plot method for \code{summarizedValidation} class objects}
\usage{
\method{plot}{summarizedValidation}(x, y = NULL, ...)
}
\description{
Plot method for \code{summarizedValidation} class objects
}

