% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{print.summarizedValidation}
\alias{print.summarizedValidation}
\title{Print method for \code{summarizedValidation} class objects}
\usage{
\method{print}{summarizedValidation}(x, ...)
}
\description{
Print method for \code{summarizedValidation} class objects
}

