% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateIndexes.R
\name{saveResampleIndexes}
\alias{saveResampleIndexes}
\title{Save a resample_indexes object into txt files.}
\usage{
saveResampleIndexes(resample_indexes, folder_path, optional_name = NULL)
}
\description{
Save a resample_indexes object into txt files.
}

