% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{summarizeValidation.multiClass}
\alias{summarizeValidation.multiClass}
\title{Summarize results from the validation step}
\usage{
\method{summarizeValidation}{multiClass}(fitted_model, ...)
}
\description{
Summarize results from the validation step
}

