% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summarizeValidation}
\alias{summarizeValidation}
\title{Generic function to summarize validation results}
\usage{
summarizeValidation(fitted_model, ...)
}
\description{
Generic function to summarize validation results
}

