% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summarizeValidationMetric}
\alias{summarizeValidationMetric}
\title{Summarize Validation metric}
\usage{
summarizeValidationMetric(validation_scores, previous_summary = NULL, ...)
}
\description{
Summarize Validation metric
}

