% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{summary.summarizedValidation}
\alias{summary.summarizedValidation}
\title{Summary method for \code{summarizedValidation} class objects}
\usage{
\method{summary}{summarizedValidation}(object, ...)
}
\description{
Summary method for \code{summarizedValidation} class objects
}

