\name{checkTags}
\alias{checkTags}
\title{Check match of the resample indexes and prediction tags}
\usage{
checkTags(resample_tag, prediction_tag, ignore_tag)
}
\value{
In case there is a mismatch it will give you a warning if
\code{ignore_tag = TRUE} or trigger an \code{stop} in case
\code{ignore_tag = FALSE}.
}
\description{
Check match of the resample indexes and prediction tags
}

