% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluateProbClass}
\alias{evaluateProbClass}
\title{Evaluate predicted class probabilities for a given score function.}
\usage{
evaluateProbClass(pred_obj, resample_indexes, type, ignore_tag = FALSE, ...)
}
\description{
Evaluate predicted class probabilities for a given score function.
}

