\name{multiClass}
\alias{multiClass}
\title{Constructor for \code{multiClass} objects}
\usage{
multiClass(resample_indexes, prob, summary_validation = NULL,
  metric_name = NULL)
}
\arguments{
  \item{resample_indexes}{S3 object of class
  \code{datasetResample} with information about Class
  Labels and the data used using for training and
  predictions. Usually generated by the
  \code{generateTestIndexes} function.}

  \item{prob}{List where each element contains a numeric
  vector with predictions for a given test set. The numeric
  vector has the form c(predictions_class1,
  predictions_class2, ..., predictions_classK) for
  classification with K classes.}

  \item{summary_validation}{Object of class
  \code{summaryValidation}, usually created by
  \code{SummaryValidationMetric} function. Default to NULL,
  when no validation step is necessary.}

  \item{metric_name}{Character vector indicating the name
  of the metric used in the validation step. Default to
  NULL, when no validation step is necessary. See the help
  of \code{evaluateProbClass} function to which metrics are
  currently available.}
}
\value{
S3 object of class \code{multiClass}.
}
\description{
\code{multiClass} objects hold final test set predictions
performed by models.
}
\details{
\code{multiClass} objects hold predictions performed by a
model over a number of test sets generated according to
some resampling of the data. A tag exist to connect the
predictions with the object that contains information about
the selected test sets. This allows results from different
models to be compared since two \code{multiClass} objects
with the same tag are guaranteed to be generated from the
same test set points.
}
\seealso{
\code{\link{generateTestIndexes}},
\code{\link{summaryValidation}},
\code{\link{evaluateProbClass}}

Other tgmMultiClass.constructors:
\code{\link{multiClassValidation}}
}

