% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{summarizeValidationMetric.multiClassValidationScores}
\alias{summarizeValidationMetric.multiClassValidationScores}
\title{Summarize Validation metric}
\usage{

  \method{summarizeValidationMetric}{multiClassValidationScores}(validation_scores,
  previous_summary = NULL, summarize_function = function(x) mean(x, na.rm =
  TRUE), ...)
}
\description{
Summarize Validation metric
}

