\name{summarizeValidationMetric.multiClassValidationScores}
\alias{summarizeValidationMetric.multiClassValidationScores}
\title{Summarize Validation metric}
\usage{

  \method{summarizeValidationMetric}{multiClassValidationScores}(validation_scores,
  previous_summary = NULL, ...)
}
\description{
Summarize Validation metric
}

